Add manual overrides to schedule e script
Zbur missed a decimal place in one of his travel sponsors, wildly inflating that group's total (and his). The other legislators who reported the same trip had the right amount but with the last two numbers representing cents. We'll run into cases where there's just a straight up issue on the form that we want to account for in the outputted data. This adds a potential way to do that to Schedule E